requirements.txt
playwright
python-telegram-bot==20.4
nest_asyncio
requests